feat: add AuthGenerateTokenUseCase for token generation and update verification code handling

+++ internal/modules/identity/repository/verification_code_repository.go
@@ -12,7 +12,7 @@ import (
 )
 
 type VerificationCodeRepository interface {
-	FindByUserAndCode(ctx context.Context, userID uint, code string) (model.VerificationCodeModel, error)
+	FindByUserAndCode(ctx context.Context, userID uint64, code string) (model.VerificationCodeModel, error)
 	Create(ctx context.Context, v model.VerificationCodeModel) (model.VerificationCodeModel, error)
 	Update(ctx context.Context, v model.VerificationCodeModel) error
 	DeleteByUserID(ctx context.Context, userID uint64) error
@@ -26,7 +26,7 @@ func NewVerificationCodeRepository(db *database.PingoDB) VerificationCodeReposit
 	return &verificationCodeRepository{db}
 }
 
-func (r *verificationCodeRepository) FindByUserAndCode(ctx context.Context, userID uint, code string) (model.VerificationCodeModel, error) {
+func (r *verificationCodeRepository) FindByUserAndCode(ctx context.Context, userID uint64, code string) (model.VerificationCodeModel, error) {
 	now := time.Now()
 	v, err := gorm.G[model.VerificationCodeModel](r.DB).
 		Where("user_id = ?", userID).
